refactor: section management 페이지 기능별로 파일 나눔
