+++ src/controller/section-management/init.js
@@ -1,0 +1,7 @@
+import { elementIds } from '../../utils.js';
+
+export const initSectionInput = (message) => {
+  alert(message);
+  elementIds.sectionOrderInput.value = '';
+  elementIds.sectionOrderInput.focus();
+};
