+++ src/controller/section-management/select-section.js
@@ -1,0 +1,7 @@
+import { renderSelect, renderSelectionRegister } from './view.js';
+
+export const selectSectionLineButton = (e) => {
+  const lineName = e.currentTarget.dataset.line;
+  renderSelectionRegister(lineName);
+  renderSelect(lineName);
+};
